Check if intermediate cert needs to be renewed.
Fixes #6542.

Right now the intermediate cert in secondary datacenters can expire.
This PR adds a check that renews the cert if it is half way through its
validity period.

In order to be able to test these changes, a new configuration option
was added: `IntermediateCertTTL` which is set extremely low in the
tests.

Todo:

* [ ] what would be a good lower bound for `IntermediateCertTTL`? Maybe
1 month?
* [ ] expose new config option everywhere and add docs
* [ ] find a way to unflake the tests. Blocking queries run for a max of
600sec and then they return. In reality that is way shorter than any
intermediate cert TTL and should be enough to trigger the expiration
check. The test however possibly fails because the check is not
triggered. One way to fix it would be to add a mechanism on the primary
server to return in intervals like 1/10ths of the intermediate cert TTL
which would trigger the check. But that would be only for tests since it
shouldn't be an issue in reality.